api-gateway/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
httpx==0.28.1
python-jose[cryptography]==3.3.0
PyJWT==2.8.0
python-multipart==0.0.6
redis==5.0.1
prometheus-client==0.19.0
python-json-logger==2.0.7
aioredis==2.0.1 
loguru>=0.6.0